Moved files to match diagram
