+++ app/src/main/java/ayds/newyork/songinfo/moredetails/model/data/ArtistRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package ayds.newyork.songinfo.moredetails.model.data.repository
+package ayds.newyork.songinfo.moredetails.model.data
 
 import ayds.newyork.songinfo.moredetails.model.domain.repository.ArtistRepository
