+++ app/src/main/java/ayds/newyork/songinfo/moredetails/presentation/view/OtherInfoWindow.kt
@@ -11,8 +11,8 @@ import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import ayds.newyork.songinfo.R
 import ayds.newyork.songinfo.moredetails.model.domain.entities.ArtistInfo
-import ayds.newyork.songinfo.moredetails.model.data.repository.external.nytimes.NYTimesAPI
-import ayds.newyork.songinfo.moredetails.model.data.repository.local.DataBase
+import ayds.newyork.songinfo.moredetails.model.data.external.nytimes.NYTimesAPI
+import ayds.newyork.songinfo.moredetails.model.data.local.nytimes.DataBase
 import ayds.newyork.songinfo.moredetails.presentation.presenter.OtherInfoUiEvent
 import ayds.newyork.songinfo.moredetails.presentation.presenter.OtherInfoUiState
 import ayds.newyork.songinfo.utils.UtilsInjector
